+++ onnx/test/automatic_upgrade_test.py
@@ -1351,6 +1351,9 @@ class TestAutomaticUpgrade(unittest.TestCase):
             attrs={"consumed_inputs": [0]},
         )
 
+    def test_SVD(self) -> None:
+        self._test_op_upgrade("SVD", 18, [[3, 4]], [[3], [3, 3], [4, 4]])
+
     def test_Tanh(self) -> None:
         self._test_op_upgrade("Tanh", 1, attrs={"consumed_inputs": [0]})